fix: resolve comments for PR #5
