+++ src/components/Charts/CBarChart/customLabel.tsx
@@ -1,28 +1,20 @@
-import { CBarChartType } from '.';
-
 import BN from '@/utils/BN';
 
-const CustomLabel = (props: any, data: CBarChartType[], activeIndex: number) => {
-  const { x, y, width, value, index } = props;
-  const maxRevenue = Math.max(...data.map((item) => item.amount));
+const CustomLabel = ({ x, y, width, value, maxRevenue }: any) => {
+  const percentage = new BN(value).div(maxRevenue).times(100).toFixed(0) + '%';
 
-  if (index === activeIndex) {
-    const percentage = new BN(value).div(maxRevenue).times(100).toFixed(0) + '%';
-
-    return (
-      <text
-        x={x + width / 2}
-        y={y - 10}
-        fill="#00C17E"
-        fontWeight="normal"
-        textAnchor="middle"
-        fontSize={13}
-      >
-        {percentage}
-      </text>
-    );
-  }
-  return null;
+  return (
+    <text
+      x={x + width / 2}
+      y={y - 10}
+      fill="#00C17E"
+      fontWeight="normal"
+      textAnchor="middle"
+      fontSize={13}
+    >
+      {percentage}
+    </text>
+  );
 };
 
 export default CustomLabel;
