+++ src/components/Charts/CBarChart/index.tsx
@@ -9,17 +9,15 @@ import {
   Rectangle,
   ResponsiveContainer,
   Tooltip,
+  TooltipProps,
   XAxis,
 } from 'recharts';
 
-import customLabel from './CustomLabel';
-import customAxisTick from './CustomAxisTick';
-import CustomTooltip from './CustomTooltip';
+import { CBarChartType } from '@/models';
 
-export type CBarChartType = {
-  name: string;
-  amount: number;
-};
+import CustomLabel from './CustomLabel';
+import CustomAxisTick from './CustomAxisTick';
+import CustomTooltip, { CustomTooltipProps } from './CustomTooltip';
 
 const data: CBarChartType[] = [
   { name: 'Jan', amount: 4000 },
@@ -46,18 +44,28 @@ const CBarChart = () => {
     [setActiveIndex],
   );
 
+  const renderCustomTooltip = (props: TooltipProps<number, string>) => {
+    const { active, payload } = props;
+    if (active && payload && payload.length) {
+      return <CustomTooltip payload={payload as CustomTooltipProps['payload']} />;
+    }
+    return null;
+  };
+
+  const maxRevenue = Math.max(...data.map((item) => item.amount));
+
   return (
     <div style={{ width: '100%', height: '100%' }}>
       <ResponsiveContainer minHeight={230} width="100%" height="100%">
         <BarChart data={data} className="-mt-3" height={230}>
           <XAxis
             dataKey="name"
-            tick={(props) => customAxisTick(props, activeIndex)}
+            tick={(props) => CustomAxisTick(props, activeIndex)}
             stroke="#D0D5DD"
             axisLine={false}
             tickLine={false}
           />
-          <Tooltip trigger="hover" cursor={false} content={<CustomTooltip />} />
+          <Tooltip trigger="hover" cursor={false} content={renderCustomTooltip} />
 
           <Bar
             dataKey="amount"
@@ -75,7 +83,11 @@ const CBarChart = () => {
             ))}
             <LabelList
               dataKey="amount"
-              content={(props) => customLabel(props, data, activeIndex)}
+              content={(props) =>
+                props.index === activeIndex ? (
+                  <CustomLabel {...props} maxRevenue={maxRevenue} />
+                ) : null
+              }
             />
           </Bar>
         </BarChart>
